requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Real-time trend data APIs
pytrends>=4.9.0  # Google Trends
tweepy>=4.14.0   # Twitter API (optional)

# AI/LLM dependencies (optional)
openai>=0.27.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# Web scraping dependencies
selenium>=4.10.0
scrapy>=2.9.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# API integrations
python-dotenv>=1.0.0
pydantic>=2.0.0
